requirements.txt
# Web Interface & Server
gradio==5.49.1
python-multipart
uvicorn
flask

# Google Cloud Services (KI & Voice)
google-cloud-texttospeech
google-genai
google-cloud-aiplatform

# Datenbank & Sicherheit
sqlalchemy
pymysql
python-dotenv
sshtunnel
paramiko==2.11.0
cryptography
argon2-cffi
passlib

# Audioverarbeitung
pydub
ffmpeg-python
ffmpy

# Quellen-Extraktion (PDF & Web)
PyPDF2
beautifulsoup4
lxml
requests
dnspython

# Hilfspakete
numpy
nltk
pytest
pytest-mock
#PYTHONPATH=. pytest --cov=services
pytest-cov
